+++ beaker-vue/src/components/chat-interface/AgentQuery.vue
