+++ beaker-vue/src/components/chat-interface/ChatPanel.vue
