+++ beaker-vue/src/components/chat-interface/DarkModeButton.vue
