+++ beaker-vue/src/components/chat-interface/ResetButton.vue
