+++ beaker-vue/src/components/chat-interface/VerticalToolbar.vue
